requirements.txt
aiogram==3.12.0
aiohttp>=3.9.0
python-dotenv==1.0.1
openai>=1.30.0
pyzbar>=0.1.9
Pillow>=10.0.0
gspread>=6.0.0
google-auth>=2.0.0
